feat(index): adding device enumeration tool
the way it works is very simple, it will request for permissions from
the user to access the various devices reachable from the tablet, then
call the `enumerateDevice()` function.

starting from Chrome 77, the enumerateDevices() function will return empty labels
and IDs until the user grants permission to access the media devices.
for this reason, we have to manually request them, but this is fine
because user has physical access in this scenario.